Added 'patience' mode, added modes to theme unlocks, added unlock text to modes
In patience mode, the obstacles only move when the screen is tapped. The
number of plays is now stored seperately for each mode (fixed
localstorage to work with this). Theme unlocking takes into account mode
as well. Modes now have unlock text, but don't yet require unlocking to
be played.